+++ src/account/types/mod.rs
@@ -6,6 +6,8 @@ pub(crate) mod address;
 /// Custom de/serialization for [`address::AddressWrapper`]
 pub(crate) mod address_serde;
 pub(crate) mod balance;
+#[cfg(feature = "participation")]
+pub mod participation;
 
 use std::str::FromStr;