+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Default` to `FilterOptions`;
+- `ParticipationEventRegistrationOptions` type;
 
 ### Changed
 
@@ -32,6 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `StorageManager::get_account_manager_data` return type changed to `Result<Option<AccountManagerBuilder>>`;
 - `Features::{tag, metadata}` encoded as hex strings instead of utf-8 strings for NFT Outputs;
 - Added `#[serde(rename = "transactionOptions")]` to `PrepareOutput::transaction_options`;
+- `AccountMethod::RegisterParticipationEvent` renamed to `AccountMethod::RegisterParticipationEvents`,
+  contains a `ParticipationEventRegistrationOptions` and expects a `ParticipationEvents` response;
+- `AccountMethod::GetParticipationEventIds` now also contains a `Node`;
 
 ### Removed